exception constants moved to root package
